追加アプリ/supabase-streamlit/requirements.txt
streamlit
supabase
uuid
python-dotenv